do not bother checking number of rows from db
The main problem is that ->rows doesn't return something useful
in our test case.  One is really not supposed to use it after
a SELECT.  Instead, we just get rows.  It will work the same:
if there was an error, RaiseError should've made it fatal.  If
there were no data, it will return nothing the first time.