ensure that Perl6 dists go into a Perl6 sub directory
We do this to not mess with Perl5 distributions, so that smokers can easily
filter distributions by path to not smoke Perl6 dists accidently.

Note that this means that the target directory one can specify when uploading
a dist is ignored for Perl6.
The tarball will always go to A/AU/AUTHOR/Perl6/<filename>.