Give full indexing report even if dist name perm fails
This commit makes a distribution name permission failure merely add an
informative header to the message, but otherwise allows the ordinary
full content of the report.  However, any "indexed" modules are fixed up
to show that they are not indexed.

This should allow better diagnostic information when dist name problems
occur.