Ensure first come and comaint come from main package on new packages.
This changes the behaviour of new packages permissions so that rather
than giving the uploader first come, it looks to the main package
for the permissions to use.  It copies both the first come and
co maintainer users to the new package.

This should mean that for all new modules uploaded permissions should
remain fairly consistent.