Expose ustatus if it's nologin or delete
so that a third party (metacpan, cpants, etc) can use 00whois.xml to find a pause_id is deleted or banned.

We usually want to hide distributions by a spammer, and mark a deleted user to hide their distributions or just to indicate the author is gone, but currently we don't have any measure to determine a pause_id is banned or deleted (except for rsynching moddata).