Revert "Bump version of operating model to 3" and "Expand operating model list of prohibitions"
This reverts commit 521a092968ff8106f4d46253243247786aa9d25a and 83d65cf33cac3e2048e1a6cfeb98cc70cfc7d12a.

after having slept over it I think it wasn't a good idea to accept
this PR. Blame the weather, blame my ignorance, I don't know. I wasn't
in the best shape and decided quickly instead of thoroughly.

I think we really should kill the catch all part of the pull request.
The Operating Model at its very heart had the idea that we guarantee
our users that we do not interfere with their uploads unless in
narrowly defined special cases.

Now we had an incident that showed us we had forgotten to talk about
security matters. We should react to this incident with a new version
that introduces security matters. No more, no less. When we have
another case that shows us that we missed something, we should seek
how to codify additional rules. That's what the spirit of the
Operating Model kind of promises.

I've now reverted the last two commits and would ask for a second
attempt at a rephrasing, this time strictly limited to what the
security incident taught us.

Thanks && and sorry for my lapse