Tests and proposed fix for #348, where first-come can be lost
If a package has one author with first-come on a package
(but only with an entry in primeur, not in perms),
and someone else has co-maint (entry in perms).
If the co-maint does the next release, then the entry in primeur
is getting deleted, so it ends up with no first-come.