on a very small (or large) version, do not index
Really, this should become a useful user message, but it's painfully
difficult to get that error up through the call stack.  This code needs
(in my opinion) to be rearchitected to make this sort of thing simple.

In the meantime, the code now detects this case specifically, so we can
do something about it later.