email testing: respect EMAIL_SENDER_TRANSPORT when testing
This updates how we use KeepDeliveries to get our default from the
environment, so that we usually send just to Test, but can run the tests
with Maildir in the env.  The tests will still pass, but we can see all
the generated mail.

This also requires updating the existing tests to not redundantly set
the EMAIL_SENDER_TRANSPORT env var.