'package' column is now case insensitive
This avoids deadlocks that are very easy to trigger right now.

The problem is that to update or delete a row related to a
package, we must do:

    where LOWER(package) = LOWER(?)

And what this means is that mysql *cannot* use an index to resolve
the query, so it scans the *entire table*.

This is slow, and also causes lock waits in simple situations:

  1. You have two processes, A1 and B1
  2. A1 begins a transaction and does:

    INSERT INTO primeur VALUES('pac::kage', 'testuser');

  3. B1 does, *outside of a transaction*:

    DELETE FROM primeur WHERE LOWER(package) = LOWER('foo::bar');

At this point, B1 is stuck waiting on a lock!

By switching to using case insensitive keys, we can do index lookups:

    DELETE FROM primeur WHERE package = 'foo::BaR';

and no weird lock waits occur, and we avoid the deadlocks.

Case sensitive collation on the 'package' column made sense back
when we imagined a world where different case files had different
meanings, but we decided against that more recently, and this will
help further enforce that.