mldistwatch: Rewrite perms/primeur in the transaction
Otherwise, we could (historically) deadlock when trying to
delete from primeur *after* we've already deleted from perms,
and folks would lose comaint permissions, or other weird cases
could happen

By bringing this logic into the transaction, we should avoid
weird broken indexing.

Surely not doing this in a transaction was a mistake when this
code was introduced.