replace use of index_status with context in reporting
In the previous commit, we'd replaced all use of ->index_status to set
the current status, and instead used the context methods to record
package status.  Still, though, those methods set *both* the new status
and the old status, and we only read things from the old status.

In this commit, we switch to only reading things from the new status,
and delete the attributes used to store the old status.