PAUSE::Config: merge new values from bootstrap
A lot of things were configured in bootstrap because the primary PAUSE
was running with the old filesystem layout.

I want to keep the default config matching the default behavior of both
test and real PAUSE instances.  The bootstrap program should be able to
build on default, rather than old settings from years ago.