raise max upload size
nginx defaults to 1mb client_max_body_size which is very limiting for
the large POSTs that PAUSE requires.

Raise to 250mb, which is a little larger than the largest files.

Note: an upload of a 200mb file failed, but in Plack/mojo, not nginx
as before. An upload of 25mb (bigger than the average perl5 release)
was fine.