TestPAUSE: set an ADMIN and CONTACT_ADDRESS while testing
ADMIN is the address that many emails are sent from.  Now that the code
will use this setting to make address objects, leaving it undef will not
do!  We get warnings.

CONTACT_ADDRESS is the setting where we usually put "modules@perl.org",
where we tell people to send the admins mail.  We will want to check
this in PAUSE daemon tests.