indexer: in one place, use proper email header formatting
This is a test run for things to come.  In most places, we only use the
user email address, not their name.  But in some places, most especially
in paused, we use their name.

There were two problems there:

1. we were using the ASCII name, which we really don't need to do
2. we were not escaping any quotes (or anything else) in the name field,
   which could cause problems if someone puts a double quote in their
   fullname

It wasn't trivial to test and fix those locations because we don't
(yet!) have any tests for paused.  This introduces the use of fullname
to mldistwatch, using the new form: get the PAUSE::MailAddress for the
user, then use its ->email_header_object method to create a
Email::MIME::Header::AddressList.  We'll switch all the mailing to use
this, soon.