pause_2017: use email header objects when possible
I don't know this code well, so this may be a bit of a wreck.

First, `$mgr->prepare_sendto` is updated.

1. prefer fullname to asciiname, because it is more polite
2. use real email header generation, not just concat
3. if we had "Bob <x@y>" and "Alice <x@y>" we now only send one email,
   not one per unique phrase; I think this is fine

Where the prepare_sendto code was basically duplicated, I have switched
things to use prepare_sendto.

Internal report emails now go to the internal reporting address, not the
public contact address.

In some places, the code was quite tough for me to read, so I just left
`$email` and the like used.  This should work fine.  It would be nice to
add phases later.

**I didn't understand this**: `$pause->{send_to}`.  That was being set
to a string, but I don't know really why or how it was used, although I
saw some hints.  The new code would've put ARRAY(0x1234) in that, so I
changed the text.  I will seek advice.