Remove indexscripts cronjob and mentions.
It has not been touched maintained in over a decade.
For the past few years, at most 1 .pl file a year has been uploaded and indexed.  The new way is to use the App:: namespace.
It has hardcoded paths to /home/kstar, which no longer exists.

Also note that lib/PAUSE.pod is also 10+ years out of date.