tests: add assert_index_updated and assert_index_not_updated
The previous method for checking this was generic, which is nice, but
sort of verbose to use.  Worse, though, I don't think it was reliable.
That code, ->file_updated_ok and ->file_not_updated_ok, worked by
keeping a cache of file stats -- but it didn't precompute those stats,
it computed them on demand.  So given this program:

  $pause = PAUSE::TestPause->new;

  do_stuff();
  my $r1 = $pause->test_reindex;
  do_other_stuff();
  $pause->file_updated_ok("A");
  my $r2 = $pause->test_reindex;
  do_more_stuff();
  my $r3 = $pause->test_reindex;

Calling `->file_updated_ok("A")` at the end would test since the
previous call.  Calling `->file_updated_ok("B")` at the end would test
starting from nothing.

In reality, we only use this method for tracking 02packages, so I've
added specific tracking of that file, with the check put onto the Result
object, considering only what changed in the `test_reindex` call that
created that Result.